osd: shuffle the ECCommon
